Extract component to dedicated file.
